+++ ci/downgrades/rector-downgrade-code-extension-demo.php
@@ -2,23 +2,13 @@
 
 declare(strict_types=1);
 
-use Rector\Core\Configuration\Option;
+use PoP\PoP\Config\Rector\Downgrade\Configurators\ExtensionDemoContainerConfigurationService;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
-require_once __DIR__ . '/rector-downgrade-code-shared.php';
-
 return static function (ContainerConfigurator $containerConfigurator): void {
-    // Shared configuration
-    doCommonContainerConfiguration($containerConfigurator);
-
-    // get parameters
-    $parameters = $containerConfigurator->parameters();
-
-    // files to skip downgrading
-    $parameters->set(Option::SKIP, [
-        // Skip tests
-        '*/tests/*',
-        '*/test/*',
-        '*/Test/*',
-    ]);
+    $containerConfigurationService = new ExtensionDemoContainerConfigurationService(
+        $containerConfigurator,
+        dirname(__DIR__, 2)
+    );
+    $containerConfigurationService->configureContainer();
 };
