+++ ci/downgrades/rector-downgrade-code.php
@@ -2,13 +2,11 @@
 
 declare(strict_types=1);
 
-use PoP\PoP\Config\Rector\Downgrade\Configurators\ContainerConfigurationService;
+use PoP\PoP\Config\Rector\Downgrade\Configurators\MonorepoContainerConfigurationService;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
-require_once __DIR__ . '/rector-downgrade-code-shared.php';
-
 return static function (ContainerConfigurator $containerConfigurator): void {
-    $containerConfigurationService = new ContainerConfigurationService(
+    $containerConfigurationService = new MonorepoContainerConfigurationService(
         $containerConfigurator,
         dirname(__DIR__, 2)
     );
